user auth: login, logout and some user specific tasks

+++ base/views.py
@@ -1,7 +1,40 @@
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib.auth.models import User
+
+def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username OR Password does not match')
+
+    context = {}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('login')
 
 def home(request):
     q = request.GET.get('q','')
@@ -11,9 +44,12 @@ def home(request):
         Q(name__icontains=q)
     )
     topics = Topic.objects.all()
+    room_count = rooms.count()
+
     context = {
         'rooms': rooms,
         'topics': topics,
+        'room_count': room_count
     }
     return render(request, 'base/home.html', context)
 
@@ -22,6 +58,7 @@ def room(request, pk):
     context = {"room":room}
     return render(request, 'base/room.html',context)
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -33,9 +70,13 @@ def create_room(request):
     context = {"form":form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room) # so the form will be prefilled with some values
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room) # pass the instancce, otherwise it will create a new room
@@ -46,8 +87,13 @@ def update_room(request, pk):
     context = {"form":form}
     return render(request, 'base/room_form.html',context)
 
+@login_required(login_url=login)
 def delete_room(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
